refactor: Update namespaces and inherit from ComponentBase in dashboard components
