+++ server/ExpertBridge.Admin/Components/Pages/DeletedContent/DeletedPosts.razor.cs
@@ -9,7 +9,7 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Caching.Hybrid;
 using Radzen;
 
-namespace ExpertBridge.Admin.Components.Pages;
+namespace ExpertBridge.Admin.Components.Pages.DeletedContent;
 
 public sealed partial class DeletedPosts : ComponentBase
 {
